+++ lib/Service/RecipeService.php
@@ -14,6 +14,8 @@ use OCP\Files\Folder;
 use OCA\Cookbook\Db\RecipeDb;
 use OCP\PreConditionNotMetException;
 use Psr\Log\LoggerInterface;
+use OCA\Cookbook\Exception\HtmlParsingException;
+use OCA\Cookbook\Exception\ImportException;
 
 /**
  * Main service class for the cookbook app.
@@ -27,14 +29,29 @@ class RecipeService {
 	private $config;
 	private $il10n;
 	private $logger;
+	
+	/**
+	 * @var HtmlDownloadService
+	 */
+	private $htmlDownloadService;
+	
+	/**
+	 * @var RecipeExtractionService
+	 */
+	private $recipeExtractionService;
 
-	public function __construct(?string $UserId, IRootFolder $root, RecipeDb $db, IConfig $config, IL10N $il10n, LoggerInterface $logger) {
+	public function __construct(
+		?string $UserId, IRootFolder $root, RecipeDb $db, IConfig $config,
+		IL10N $il10n, LoggerInterface $logger, HtmlDownloadService $downloadService,
+		RecipeExtractionService $extractionService) {
 		$this->user_id = $UserId;
 		$this->root = $root;
 		$this->db = $db;
 		$this->config = $config;
 		$this->il10n = $il10n;
 		$this->logger = $logger;
+		$this->htmlDownloadService = $downloadService;
+		$this->recipeExtractionService = $extractionService;
 	}
 
 	/**
@@ -417,6 +434,7 @@ class RecipeService {
 	 * @param string $html
 	 *
 	 * @return array
+	 * @deprecated
 	 */
 	private function parseRecipeHtml($url, $html) {
 		if (!$html) {
@@ -799,36 +817,27 @@ class RecipeService {
 	}
 
 	/**
-	 * @param string $url
+	 * Download a recipe from a url and store it in the files
 	 *
+	 * @param string $url The recipe URL
+	 * @throws Exception
 	 * @return File
 	 */
-	public function downloadRecipe($url) {
-		$host = parse_url($url);
-
-		if (!$host) {
-			throw new Exception('Could not parse URL');
+	public function downloadRecipe(string $url): File {
+		$this->htmlDownloadService->downloadRecipe($url);
+		
+		try {
+			$json = $this->recipeExtractionService->parse($this->htmlDownloadService->getDom());
+		} catch (HtmlParsingException $ex) {
+			throw new ImportException($ex->getMessage(), null, $ex);
 		}
-
-		$opts = [
-			"http" => [
-				"method" => "GET",
-				"header" => "User-Agent: Nextcloud Cookbook App"
-			]
-		];
-
-		$context = stream_context_create($opts);
-
-		$html = file_get_contents($url, false, $context);
-
-		if (!$html) {
-			throw new Exception('Could not fetch site ' . $url);
-		}
-
-		$json = $this->parseRecipeHtml($url, $html);
-
+		
+		$json = $this->checkRecipe($json);
+		
 		if (!$json) {
-			throw new Exception('No recipe data found');
+			$this->logger->error('Importing parsers resulted in null recipe. ' .
+				'This is most probably a bug. Please report.');
+			throw new ImportException($this->il10n->t('No recipe data found. This is a bug'));
 		}
 
 		$json['url'] = $url;